% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ergodic_sim}
\alias{ergodic_sim}
\title{ergodic process simulator}
\usage{
ergodic_sim(agents, rounds, money, win, loss, prob)
}
\arguments{
\item{agents}{integer specifying the number of agents}

\item{rounds}{integer specifying the number of rounds}

\item{money}{integer specifying initial money for agents}

\item{win}{double specifying multiplier for money given a win (should be strictly positive)}

\item{loss}{double specifying multiplier for money given a loss (should be strictly positive)}

\item{prob}{double specifying probability of a win}
}
\value{
a matrix of agent money per round where rows are rounds and columns are agents
}
\description{
ergodic process simulator
}
