% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_gd.R
\name{gen_data}
\alias{gen_data}
\title{data generator function}
\usage{
gen_data(n = 1000, n_features = 10)
}
\arguments{
\item{n}{int specifying number of observations to generate}

\item{n_features}{int specifying number of variables to generate}
}
\value{
a data.frame of variables
}
\description{
data generator function
}
