% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_mean.R
\name{k_mean}
\alias{k_mean}
\title{k-mean clustering algorithm}
\usage{
k_mean(points, k, max_iter = 500L, n_init = 5L)
}
\arguments{
\item{points}{a numeric m x n matrix with m observations defined by n coordinate values.}

\item{k}{integer, number of clusters}

\item{max_iter}{integer, maximum number of clustering optimization iterations}

\item{n_init}{integer, number of random initial cluster starting points to try}
}
\value{
a list of squared errors, cluster assignments and cluster centroid coordinates for each iteration
}
\description{
k-mean clustering algorithm
}
