% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmean}
\alias{kmean}
\title{kmean}
\usage{
kmean(points, k, max_iter)
}
\arguments{
\item{points}{a matrix where rows are individual points and columns are their coordinates}

\item{k}{integer specifying the number of clusters to search for}

\item{max_iter}{integer specifying the maximum number of iterations to run the algorithm for}
}
\value{
a list with an integer vector specifying the cluster assignment of points and a matrix with centroid point coordinates for each cluster
}
\description{
kmean
}
