% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_gd.R
\name{lm_gd}
\alias{lm_gd}
\title{linear regression with adaptive gradient descent}
\usage{
lm_gd(data, y, x, lr = 0.01, iter = 1000)
}
\arguments{
\item{data}{a data.frame or matrix of data}

\item{y}{a vector of the column name or index of the response variable}

\item{x}{a vector of column names or indeces of the predictor variables}

\item{lr}{dbl specifying learning rate}

\item{iter}{dbl specifying number of iterations}
}
\value{
a list including a named vector of estimated coefficients, a vector of intercept values for each iteration and a matrix of coefficients for each iteration.
}
\description{
linear regression with adaptive gradient descent
}
