% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lt_permute}
\alias{lt_permute}
\title{helper to permute sampling data}
\usage{
lt_permute(link_list, wave, name)
}
\arguments{
\item{link_list}{List holding indices of linked units for each unit}

\item{wave}{integer vector holding the sampling wave of each unit}

\item{name}{integer vector holding the name of each unit}
}
\value{
vector of integer vectors holding permuted sampling waves
}
\description{
helper to permute sampling data
}
\keyword{internal}
