% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pareto_sim}
\alias{pareto_sim}
\title{pareto process simulator}
\usage{
pareto_sim(pop_size, mon, prop, iter)
}
\arguments{
\item{pop_size}{integer specifying the number of agents}

\item{mon}{integer vector specifying initial money for each agent}

\item{prop}{double vector specifying the win probability for each agent}

\item{iter}{integer specifying the number of iterations to run}
}
\value{
a list of agent money vectors
}
\description{
pareto process simulator
}
