% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ruin_sim}
\alias{ruin_sim}
\title{gamblers ruin simulation}
\usage{
ruin_sim(
  n_runs,
  money_start,
  money_end,
  p_win,
  risk_adjust,
  adaptive,
  prior_p_win
)
}
\arguments{
\item{n_runs}{integer number of simulations}

\item{money_start}{double initial money}

\item{money_end}{double money to be achieved}

\item{p_win}{double win probability}

\item{risk_adjust}{double proportion of money to bet}

\item{adaptive}{bool indicating whether to use adaptive betting (based on bayesian estimation of win probability)}

\item{prior_beta}{integer vector holding alpha and beta parameters of the p_win prior}
}
\value{
a List holding a vector indicating rounds and a vector indicating busts for each run
}
\description{
gamblers ruin simulation
}
